+++ internal/resources/extensibility/extensions.go
@@ -16,7 +16,27 @@
 package extensibility
 
 type Extension struct {
-	ExtensionID string `json:"extensionId"`
-	Name        string `json:"name"`
-	Description string `json:"description"`
+	ExtensionID   string            `yaml:"extensionId,omitempty" json:"extensionId,omitempty"`
+	ExtensionType string            `yaml:"extensionType,omitempty" json:"extensionType,omitempty"`
+	Name          string            `yaml:"name" json:"name"`
+	Description   string            `yaml:"description" json:"description"`
+	Entry         string            `yaml:"entry" json:"entry"`
+	Contexts      []string          `yaml:"contexts" json:"contexts"`
+	Permissions   []string          `yaml:"permissions,omitempty" json:"permissions,omitempty"`
+	Routes        []*ExtensionRoute `yaml:"routes,omitempty" json:"routes,omitempty"`
+}
+
+type Icon struct {
+	Name string `json:"name"`
+}
+
+type ExtensionRoute struct {
+	ID              string            `yaml:"id" json:"id"`
+	LocationID      string            `yaml:"locationId" json:"locationId"`
+	ParentID        string            `yaml:"parentId,omitempty" json:"parentId,omitempty"`
+	DestinationPath string            `yaml:"destinationPath,omitempty" json:"destinationPath,omitempty"`
+	LabelIntl       map[string]string `yaml:"labelIntl,omitempty" json:"labelIntl,omitempty"`
+	RenderType      string            `yaml:"renderType,omitempty" json:"renderType,omitempty"`
+	Order           *float64          `yaml:"order,omitempty" json:"order,omitempty"`
+	Icon            *Icon             `yaml:"icon,omitempty" json:"icon,omitempty"`
 }